Scheduler UT update

+++ HimalayaMain/Master/Components/Scheduler/Test/Mock/MockIDeviceProcessing.h
@@ -207,7 +207,7 @@ public:
     MOCK_METHOD0(Destroy, void());
     MOCK_METHOD1(NotifySavedServiceInfor, void(const QString& deviceType));
     MOCK_METHOD0(ResetActiveCarbonFilterLifeTime, void());
-    MOCK_METHOD1(IDSetAlarm, ReturnCode_t(int opcode));
+    MOCK_METHOD1(IDSetAlarm, ReturnCode_t(quint8 opcode));
     MOCK_METHOD2(ALControlValve, ReturnCode_t(quint8 ValveIndex, quint8 ValveState));
     MOCK_METHOD2(RVSetTemperatureSwitchState, ReturnCode_t(qint8 HeaterVoltage, qint8 AutoType));
     MOCK_METHOD3(RTSetTemperatureSwitchState, ReturnCode_t(RTTempCtrlType_t Type, qint8 HeaterVoltage, qint8 AutoType));
